Test_Bot/requirements.txt
# VK Bot Dependencies
vkbottle==4.4.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration & Utilities
python-dotenv==1.0.0
aiofiles==24.1.0
aiohttp==3.9.1
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For production deployment
gunicorn==21.2.0
uvicorn==0.24.0
